move the total logic to a hook
